Restructure slsa-provenance cli
