+++ go.mod
@@ -1,3 +1,8 @@
 module github.com/philips-labs/slsa-provenance-action
 
 go 1.17
+
+require (
+	github.com/peterbourgon/ff/v3 v3.1.2
+	github.com/pkg/errors v0.9.1
+)
